+++ src/common/mod.rs
@@ -1,2 +1,4 @@
+pub mod properties;
+
 #[cfg(feature = "async")]
 pub mod fetch;
